Create Doctrines hooks to dispatch the events
A hook will listen to Doctrines onFlush events, and will rely upon an
event tracker to stage the events that have happened during the
transaction.
A second hook will listen to the postFlush event, which is triggered
only in case of a successful flush, and will dispatch those changes via
a Symfony's event dispatcher.