Add strategy to always have ordered outbox entries
During tests with a huge volume of inserting data, we noticed that
some records were persisted on the wrong order, and it would be a huge
problem since the clients (consumers) tends to rely on the record order
to take actions.
The first approach was changing the precision of the created_at
timestamp column (to have milliseconds), but it doesn't work and results
were the same.
The second approach was adding an auto-increment column and make it
primary key, but again the problems persisted.
Finally, as the third and now effective solution, a foreign key constraint
was created on the new "previous_event" column and a reference was made
to the current primary key column (id). With this association, the
database (tests were made using Postgres) always guarantee the order of
insertions due to the relationship (FK) nature.